refactor(slots): extract npu/ package (P3-slots §1d)
Moves the NPU FLM-trio shadow reconciler out of slots/manager.py into a
new slots/npu/ package: is_npu_trio_shadow (predicate) + _TRIO_SHADOW_SPEC
+ the full reconcile_npu_trio_slots body now live in slots/npu/trio.py as
reconcile_trio_slots(mgr), taking a narrow NpuTrioHost protocol
(iter_configs/create/_invalidate_cfg_cache) instead of the concrete
SlotManager.

slots/manager.py keeps reconcile_npu_trio_slots as a one-line delegator
(api/__init__.py:911 calls it) and re-exports is_npu_trio_shadow (5 core
call sites + tests/slots/test_npu_trio_shadow.py import it directly).
slots/__init__.py also re-exports is_npu_trio_shadow per the P3-slots §5
contract.

§11.1 seam note carried into npu/trio.py's docstring: this shadow
lifecycle is interim (a future uniform id-keyed slot lifecycle dissolves
it) — kept as one importable function, not a mixin, so it deletes cleanly
later.

No behavior change. tests/slots (379) green.
